+++ mbsimgui/mbsimgui/parameter.cc
@@ -179,6 +179,7 @@ namespace MBSimGUI {
   void ImportParameter::updateValue() {
     Parameter::updateValue();
     value = MBXMLUtils::E(element)->getFirstTextChild()?QString::fromStdString(MBXMLUtils::X()%MBXMLUtils::E(element)->getFirstTextChild()->getData()):"";
+    type = E(element)->getAttribute("type");
   }
 
   Parameters::Parameters(EmbedItemData *parent) : ParameterItem(parent) {