+++ mbsimgui/mbsimgui/parameter_property_dialog.h
@@ -102,6 +102,7 @@ namespace MBSimGUI {
       xercesc::DOMElement* writeXMLFile(xercesc::DOMNode *element, xercesc::DOMNode *ref=nullptr) override;
     protected:
       ExtWidget *value;
+      ExtWidget *type;
   };
 
 }